+++ 00_gcp.sh
@@ -1,5 +1,8 @@
 # https://cloud.google.com/appengine/docs/standard/java/accessing-instance-metadata
 # https://cloud.google.com/compute/docs/reference/rest/v1/instances/list
+GCP_EXTERNAL_IP=$(curl -s -H "Metadata-Flavor: Google" "metadata.google.internal/computeMetadata/v1/instance/network-interfaces/0/access-configs/0/external-ip")
+GCP_EXTERNAL_IP=${GCP_EXTERNAL_IP:-'None'}
+
 GCP_PROJECT=$(curl -s -H "Metadata-Flavor: Google" "metadata.google.internal/computeMetadata/v1/project/project-id")
 GCP_INSTANCE_MACHINE_TYPE=$(curl -s -H "Metadata-Flavor: Google" "metadata.google.internal/computeMetadata/v1/instance/machine-type" | xargs basename)
 GCP_INSTANCE_PREEMPTIBLE=$(curl -s -H "Metadata-Flavor: Google" "metadata.google.internal/computeMetadata/v1/instance/scheduling/preemptible")
@@ -7,6 +10,7 @@ GCP_INSTANCE_VPC=$(curl -s -H "Metadata-Flavor: Google" "metadata.google.interna
 GCP_INSTANCE_ZONE=$(curl -s -H "Metadata-Flavor: Google" "metadata.google.internal/computeMetadata/v1/instance/zone" | xargs basename)
 
 echo -e "===== GCP INSTANCE METADATA ==================================================
+ ${COLOR_COLUMN}- External IP${RESET_COLORS}........: ${COLOR_VALUE} ${GCP_EXTERNAL_IP} ${RESET_COLORS}
  ${COLOR_COLUMN}- Project ID${RESET_COLORS}.........: ${COLOR_VALUE} ${GCP_PROJECT} ${RESET_COLORS}
  ${COLOR_COLUMN}- Machine Type${RESET_COLORS}.......: ${COLOR_VALUE} ${GCP_INSTANCE_MACHINE_TYPE} ${RESET_COLORS}
  ${COLOR_COLUMN}- Preemptible${RESET_COLORS}........: ${COLOR_VALUE} ${GCP_INSTANCE_PREEMPTIBLE} ${RESET_COLORS}